% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_document.R
\name{as_document}
\alias{as_document}
\title{Convert an \code{rmd_ast}, \code{rmd_tibble}, or any ast node into text.}
\usage{
as_document(x, padding = "", collapse = NULL, ...)
}
\arguments{
\item{x}{\code{rmd_ast}, \code{rmd_tibble}, or parsermd node object.}

\item{padding}{Padding to add between nodes when assembling the text.}

\item{collapse}{If not \code{NULL}, use value to collapse lines.}

\item{...}{Passed to \code{to_ast()} when converting \code{rmd_collection} or \code{qmd_collection}.}
}
\value{
Returns a character vector.
}
\description{
Convert an \code{rmd_ast}, \code{rmd_tibble}, or any ast node into text.
}
