% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{parse_rmd}
\alias{parse_rmd}
\alias{parse_qmd}
\title{Parse an R Markdown or Quarto document}
\usage{
parse_rmd(rmd)

parse_qmd(qmd)
}
\arguments{
\item{rmd}{Either the path to an \code{Rmd} file or a character vector containing the contents
of a R Markdown document.}

\item{qmd}{Either the path to an \code{qmd} file or a character vector containing the contents
of a Quarto document.}
}
\value{
Returns a \code{rmd_ast} object.
}
\description{
Documents are parsed into an \code{rmd_ast} object.
}
\examples{
parse_rmd(system.file("examples/hw01.Rmd", package="parsermd"))

}
