% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsermd-package.R
\docType{package}
\name{parsermd-package}
\alias{parsermd}
\alias{parsermd-package}
\title{parsermd: Formal Parser and Related Tools for R Markdown Documents}
\description{
An implementation of a formal grammar and parser for R Markdown documents using the Boost Spirit X3 library. It also includes a collection of high level functions for working with the resulting abstract syntax tree.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rundel.github.io/parsermd/}
  \item \url{https://github.com/rundel/parsermd}
  \item Report bugs at \url{https://github.com/rundel/parsermd/issues}
}

}
\author{
\strong{Maintainer}: Colin Rundel \email{rundel@gmail.com}

}
\keyword{internal}
