% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R, R/parsermd-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{type_sum}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{pillar}{\code{\link[pillar]{type_sum}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

