% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_ast}
\alias{rmd_ast}
\title{AST container for multiple nodes}
\usage{
rmd_ast(nodes = list())
}
\arguments{
\item{nodes}{List of rmd node objects}
}
\description{
S7 class representing a collection of R Markdown nodes
}
