% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_ast_append.R
\name{rmd_ast_append}
\alias{rmd_ast_append}
\alias{rmd_ast_prepend}
\title{Append or preprend nodes to an ast}
\usage{
rmd_ast_append(x, ...)

rmd_ast_prepend(x, ...)
}
\arguments{
\item{x}{An object containing an \code{rmd_ast} of some kind, e.g.
\code{rmd_ast}, \code{rmd_tibble}, or \code{rmd_collection}.}

\item{...}{A collections of ast nodes to append or prepend.}
}
\value{
An object of the same class as \code{x}
}
\description{
Functions for adding nodes to the beginning or end of an ast.
}
