% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_chunk}
\alias{rmd_chunk}
\title{Code chunk node}
\usage{
rmd_chunk(
  engine = "r",
  name = "",
  options = list(),
  code = character(),
  indent = "",
  n_ticks = 3L
)
}
\arguments{
\item{engine}{Character. Language engine}

\item{name}{Character. Chunk name}

\item{options}{List. Combined chunk options (traditional and YAML)}

\item{code}{Character vector. Code lines}

\item{indent}{Character. Indentation}

\item{n_ticks}{Integer. Number of backticks}
}
\description{
S7 class representing an executable code chunk
}
