% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_code_block_literal}
\alias{rmd_code_block_literal}
\title{Code block literal node}
\usage{
rmd_code_block_literal(
  attr = "",
  code = character(),
  indent = "",
  n_ticks = 3L
)
}
\arguments{
\item{attr}{Character. Raw attribute content from {{...}}}

\item{code}{Character vector. Code lines}

\item{indent}{Character. Indentation}

\item{n_ticks}{Integer. Number of backticks}
}
\description{
S7 class representing a code block with {{...}} attributes
}
