% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_create.R
\name{rmd_create}
\alias{rmd_create}
\alias{rmd_ast}
\alias{rmd_yaml}
\alias{rmd_heading}
\alias{rmd_code_block}
\alias{rmd_chunk}
\alias{rmd_raw_chunk}
\alias{rmd_fenced_div_open}
\alias{rmd_fenced_div_close}
\alias{rmd_markdown}
\alias{rmd_inline_code}
\alias{rmd_shortcode}
\title{AST node creation}
\usage{
rmd_ast(...)

rmd_yaml(...)

rmd_heading(name, level)

rmd_code_block(attr = "", code = character(), indent = "", n_ticks = 3L)

rmd_chunk(
  name = NULL,
  engine = "r",
  options = list(),
  yaml_options = list(),
  code = character(),
  indent = "",
  n_ticks = 3L
)

rmd_raw_chunk(format, code = character(), indent = "", n_ticks = 3L)

rmd_fenced_div_open(attr = character())

rmd_fenced_div_close()

rmd_markdown(lines = character())

rmd_inline_code(engine = "", code = "")

rmd_shortcode(func, args = character(), start = -1L, length = -1L)
}
\arguments{
\item{...}{Elements within the node.}

\item{name}{Character. Heading or chunk name.}

\item{level}{Integer. Heading level (1-6).}

\item{attr}{Character. Attributes for code block or fenced div.}

\item{code}{Character. Code lines for code block or chunk.}

\item{indent}{Character. Indentation for code block or chunk.}

\item{n_ticks}{Integer. Number of backticks for code block or chunk.}

\item{engine}{Character. Language engine for chunk or inline code}

\item{options}{List. Chunk options.}

\item{yaml_options}{List. Chunk yaml options.}

\item{format}{Character. Format for raw chunk.}

\item{lines}{Character. Markdown text lines.}

\item{func}{Character. Shortcode function name.}

\item{args}{Character. Shortcode arguments.}

\item{start}{Integer. Shortcode start position.}

\item{length}{Integer. Shortcode length.}
}
\value{
An object with class matching the function name, e.g. \code{rmd_ast()} returns an \code{rmd_ast} object.
}
\description{
Functions for creating ast nodes,
\itemize{
\item \code{rmd_ast()} - Create an ast container of nodes
\item \code{rmd_yaml()} - Create a yaml node
\item \code{rmd_heading()} - Create a heading node
\item \code{rmd_code_blokc()} - Create a markdown code block node
\item \code{rmd_chunk()} - Create a chunk node
\item \code{rmd_raw_chunk()} - Create a raw chunk node
\item \code{rmd_fenced_div_open()} - Create a fenced div open node
\item \code{rmd_fenced_div_close()} - Create a fenced div close node
\item \code{rmd_markdown()} - Create a markdown container node of \code{rmd_markdown_line}s
\item \code{rmd_inline_code()} - Create an inline code node
\item \code{rmd_shortcode()} - Create a shortcode node
}
}
