% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_fenced_div_open}
\alias{rmd_fenced_div_open}
\title{Opening fenced div node}
\usage{
rmd_fenced_div_open(attr = character(0))
}
\arguments{
\item{attr}{Character vector. Div attributes}
}
\description{
S7 class representing the opening of a fenced div
}
