% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_get_node.R
\name{rmd_get_node}
\alias{rmd_get_node}
\alias{rmd_get_chunk}
\alias{rmd_get_markdown}
\title{Extract a node from an rmd object}
\usage{
rmd_get_node(x, sec_refs = NULL, type_refs = NULL, name_refs = NULL, ...)

rmd_get_chunk(x, sec_refs = NULL, name_refs = NULL)

rmd_get_markdown(x, sec_refs = NULL)
}
\arguments{
\item{x}{rmd object, e.g. \code{rmd_ast} or \code{rmd_tibble}.}

\item{sec_refs}{Section references, TODO - add details.}

\item{type_refs}{Node type references, TODO - add details.}

\item{name_refs}{Name references, TODO - add details.}

\item{...}{Unused, for extensibility.}
}
\description{
Returns a single node using \link{rmd_subset} selection.
}
