% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_has_node.R
\name{rmd_has_node}
\alias{rmd_has_node}
\alias{rmd_has_chunk}
\alias{rmd_has_markdown}
\title{Determine if an rmd object has a specific node}
\usage{
rmd_has_node(x, sec_refs = NULL, type_refs = NULL, name_refs = NULL, ...)

rmd_has_chunk(x, sec_refs = NULL, name_refs = NULL, ...)

rmd_has_markdown(x, sec_refs = NULL, ...)
}
\arguments{
\item{x}{rmd object, e.g. \code{rmd_ast} or \code{rmd_tibble}.}

\item{sec_refs}{Section references, TODO - add details.}

\item{type_refs}{Node type references, TODO - add details.}

\item{name_refs}{Name references, TODO - add details.}

\item{...}{Unused, for extensibility.}
}
\description{
Returns \code{TRUE} if the a matching node exists within the rmd object.
}
