% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_inline_code}
\alias{rmd_inline_code}
\title{Inline code node}
\usage{
rmd_inline_code(
  engine = "",
  code = "",
  braced = FALSE,
  start = -1L,
  length = -1L
)
}
\arguments{
\item{engine}{Character. Language engine}

\item{code}{Character. Code content}

\item{braced}{Logical. Whether code is braced}

\item{start}{Integer. Start position}

\item{length}{Integer. Length}
}
\description{
S7 class representing inline code
}
