% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_markdown}
\alias{rmd_markdown}
\title{Markdown text node}
\usage{
rmd_markdown(lines = character(0))
}
\arguments{
\item{lines}{Character vector. Markdown text lines}
}
\description{
S7 class representing markdown text content
}
