% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tree.R
\name{rmd_node_depth}
\alias{rmd_node_depth}
\title{Calculate nesting depth for each node in an R Markdown AST}
\usage{
rmd_node_depth(x)
}
\arguments{
\item{x}{An rmd_ast object}
}
\value{
An integer vector of nesting depths (0-indexed) for each node
}
\description{
Calculate nesting depth for each node in an R Markdown AST
}
