% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_node_sections.R
\name{rmd_node_sections}
\alias{rmd_node_sections}
\title{Find the sections for each rmd object node}
\usage{
rmd_node_sections(x, levels = 1:6, drop_na = FALSE)
}
\arguments{
\item{x}{An rmd object, e.g. \code{rmd_ast} or \code{rmd_tibble}.}

\item{levels}{Limit which section heading levels to return.}

\item{drop_na}{Should \code{NA} sections be dropped.}
}
\value{
A list of section names for each node.
}
\description{
Uses the section headings of an rmd object to indentify the heirarchical
structure of the document.
}
