% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_raw_chunk}
\alias{rmd_raw_chunk}
\title{Raw chunk node}
\usage{
rmd_raw_chunk(
  format = character(0),
  code = character(),
  indent = "",
  n_ticks = 3L
)
}
\arguments{
\item{format}{Character. Output format}

\item{code}{Character vector. Code lines}

\item{indent}{Character. Indentation}

\item{n_ticks}{Integer. Number of backticks}
}
\description{
S7 class representing a raw output chunk
}
