% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_shortcode}
\alias{rmd_shortcode}
\title{Shortcode node}
\usage{
rmd_shortcode(
  func = character(0),
  args = character(),
  start = -1L,
  length = -1L
)
}
\arguments{
\item{func}{Character. Function name}

\item{args}{Character vector. Function arguments}

\item{start}{Integer. Start position}

\item{length}{Integer. Length}
}
\description{
S7 class representing a shortcode function call
}
