% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_source.R
\name{rmd_source}
\alias{rmd_source}
\title{Source the code chunks of an Rmd document}
\usage{
rmd_source(x, ..., label_comment = TRUE, use_eval = TRUE)
}
\arguments{
\item{x}{An Rmd document (e.g. \code{rmd_ast}, \code{rmd_tibble}, etc.)}

\item{...}{Additional arguments passed to \code{\link{source}}}

\item{label_comment}{Attach chunk labels as comment before each code block}

\item{use_eval}{Use the \code{eval} chunk option to determine if code is included}
}
\description{

}
\examples{

}
