% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_span}
\alias{rmd_span}
\title{Span node}
\usage{
rmd_span(
  text = "",
  id = character(),
  classes = character(),
  attr = character()
)
}
\arguments{
\item{text}{Character. Span text content}

\item{id}{Character vector. HTML ID (length 0 or 1)}

\item{classes}{Character vector. CSS classes}

\item{attr}{Named character vector. Additional attributes}
}
\description{
S7 class representing a span with attributes
}
