% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_subset.R
\name{rmd_subset}
\alias{rmd_subset}
\title{Subset the nodes of an rmd object}
\usage{
rmd_subset(
  x,
  sec_refs = NULL,
  type_refs = NULL,
  name_refs = NULL,
  exclude = FALSE,
  keep_yaml = TRUE,
  keep_setup = FALSE,
  ...
)
}
\arguments{
\item{sec_refs}{Section references, TODO - add details}

\item{type_refs}{Node type references, TODO - add details}

\item{name_refs}{Name references, TODO - add details}

\item{exclude}{Should the matching nodes be excluded}

\item{keep_yaml}{Should the document setup chunk be kept.}

\item{...}{Unused, for extensibility.}
}
\description{
Subset an rmd object based on sections, node types, or names.
}
