% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_subset_util.R
\name{rmd_subset_util}
\alias{rmd_subset_util}
\alias{rmd_get_node}
\alias{rmd_get_chunk}
\alias{rmd_get_markdown}
\alias{rmd_has_node}
\alias{rmd_has_chunk}
\alias{rmd_has_markdown}
\title{rmd_subset utility functions}
\usage{
rmd_get_node(x, sec_refs = NULL, type_refs = NULL, name_refs = NULL, ...)

rmd_get_chunk(x, sec_refs = NULL, name_refs = NULL)

rmd_get_markdown(x, sec_refs = NULL)

rmd_has_node(x, sec_refs = NULL, type_refs = NULL, name_refs = NULL, ...)

rmd_has_chunk(x, sec_refs = NULL, name_refs = NULL, ...)

rmd_has_markdown(x, sec_refs = NULL, ...)
}
\arguments{
\item{x}{rmd object, e.g. \code{rmd_ast} or \code{rmd_tibble}.}

\item{sec_refs}{Section references, TODO - add details.}

\item{type_refs}{Node type references, TODO - add details.}

\item{name_refs}{Name references, TODO - add details.}

\item{...}{Unused, for extensibility.}
}
\value{
\itemize{
\item \verb{rmd_get_*()} functions returns a single Rmd node object (e.g. \code{rmd_heading}, \code{rmd_chunk}, \code{rmd_markdown}, etc.)
\item \verb{rmd_has_*()} functions return \code{TRUE} if a matching node exists, \code{FALSE} otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Tools for selecting or checking a single node using \code{\link[=rmd_subset]{rmd_subset()}} selection.
}
