% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_document.R
\name{as_document}
\alias{as_document}
\title{Convert an parsermd \code{rmd_ast} or related object into a text document}
\usage{
as_document(obj, padding = "")
}
\arguments{
\item{obj}{\code{rmd_ast}, \code{rmd_tibble}, or parsermd node object}

\item{padding}{Padding to add between nodes when assembling the document.}
}
\description{
Convert an parsermd \code{rmd_ast} or related object into a text document
}
