% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{parse_rmd}
\alias{parse_rmd}
\title{Parse an RMarkdown document}
\usage{
parse_rmd(rmd, allow_incomplete = FALSE, parse_yaml = TRUE)
}
\arguments{
\item{rmd}{Either the path to an \code{Rmd} file or a character vector containing the contents
of a RMarkdown document.}

\item{allow_incomplete}{Allow incomplete parsing of the document.}

\item{parse_yaml}{Use the \link[yaml:read_yaml]{yaml} package to parse the document's yaml.}
}
\description{
Documents are parse into an \code{rmd_ast} object.
}
\examples{
parse_rmd(system.file("hw01.Rmd", package="parsermd"))

}
