% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R, R/check.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

