% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_fenced_div_open}
\alias{rmd_fenced_div_open}
\title{Opening fenced div node}
\usage{
rmd_fenced_div_open(
  id = character(),
  classes = character(),
  attr = character()
)
}
\arguments{
\item{id}{Character vector. HTML ID (length 0 or 1)}

\item{classes}{Character vector. CSS classes}

\item{attr}{Named character vector. Key-value attributes (keys as names)}
}
\description{
S7 class representing the opening of a fenced div
}
