% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_s7.R
\name{rmd_heading}
\alias{rmd_heading}
\title{Heading node}
\usage{
rmd_heading(name = character(0), level = integer(0))
}
\arguments{
\item{name}{Character. Heading text}

\item{level}{Integer. Heading level (1-6)}
}
\description{
S7 class representing a markdown heading
}
