% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_select_helpers.R
\name{rmd_select_helpers}
\alias{rmd_select_helpers}
\alias{has_type}
\alias{by_section}
\alias{by_label}
\title{Rmd selection helper functions}
\usage{
has_type(types)

by_section(sec_ref, keep_parents = TRUE)

by_label(label)
}
\arguments{
\item{types}{Vector of character type names, e.g. \code{rmd_chunk}, \code{rmd_heading}, etc.}

\item{sec_ref}{character vector, a section reference selector. See details below
for further details on how these are constructed.}

\item{keep_parents}{Logical, retain the parent headings of selected sections.
Default: \code{TRUE}}

\item{label}{character vector, glob patterns for matching chunk labels.}
}
\description{
These functions are used in conjunction with \code{\link[=rmd_select]{rmd_select()}} to
select nodes from an Rmd ast.
\itemize{
\item \code{has_type()} - selects all nodes that have the given type(s).
\item \code{by_section()} - uses section selectors to select nodes.
}
}
\details{
\subsection{Section reference selectors}{

Section reference selectors are a simplified version of CSS selectors that
are designed to enable the selection nodes in a way that respects the implied
hierarchy of a document's section headings.

They consist of a character vector of heading names where each subsequent value
is assumed to be nested within the preceding value. For example, the section
selector \code{c("Sec 1", "Sec 2")} would select all nodes that are contained within
a section named \verb{Sec 2} that is in turn contained within a section named \verb{Sec 1}
(or a section contained within a section named \verb{Sec 1}, and so on).

The individual section names can be specified using wildcards (aka globbing
patterns), which may match one or more sections within the document, e.g.
\code{c("Sec 1", "Sec *")}. See \code{\link[utils:glob2rx]{utils::glob2rx()}} or
\href{https://en.wikipedia.org/wiki/Glob_(programming)}{wikipedia}
for more details on the syntax for these patterns.
}
}
