% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_subset.R
\name{rmd_subset}
\alias{rmd_subset}
\title{Subset the nodes of an rmd object}
\usage{
rmd_subset(
  x,
  sec_refs = NULL,
  type_refs = NULL,
  name_refs = NULL,
  exclude = FALSE,
  keep_yaml = TRUE,
  keep_setup = FALSE,
  ...
)
}
\arguments{
\item{x}{rmd object, e.g. \code{rmd_ast} or \code{rmd_tibble}.}

\item{sec_refs}{Section references, TODO - add details.}

\item{type_refs}{Node type references, TODO - add details.}

\item{name_refs}{Name references, TODO - add details.}

\item{exclude}{Should the matching nodes be excluded.}

\item{keep_yaml}{Should the document yaml be kept.}

\item{keep_setup}{Should the document setup chunk be kept.}

\item{...}{Unused, for extensibility.}
}
\value{
Returns a subset Rmd object (either \code{rmd_ast} or \code{rmd_tibble} depending on input).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Subset an rmd object based on sections, node types, or names.
}
